GuessTheNumberProject
Console game "Guess A Number". "Guess A Number" is a game in which your opponent, "the computer" chooses a random number between "1 and 100" and your task is to guess this number. After each number you enter, the computer will give you a hint of whether the number is greater or less than the number you selected until you guess the correct number.